Reorganize celery tasks into modules
